requirements.txt
Django==1.5.1
distribute==0.6.34
dj-database-url==0.2.1
gunicorn==0.17.4
httplib2==0.8
psycopg2==2.5
python-dateutil==2.1
pytz==2013b
six==1.3.0
twilio==3.4.5
unittest2==0.5.1
wsgiref==0.1.2
django-allauth==0.10.1
django-localflavor-us==1.1
pandas==0.11.0
South==0.8.1
django-registration==1.0
django-picklefield==0.3.0